requirements.txt
absl-py==0.4.1
astor==0.7.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
ConfigArgParse==0.14.0
future==0.17.1
gast==0.2.0
grpcio==1.12.1
idna==2.8
Markdown==2.6.11
mkl-fft==1.0.6
mkl-random==1.0.1
numpy==1.15.2
olefile==0.46
Pillow==5.2.0
protobuf==3.6.0
pycparser==2.19
requests==2.22.0
scipy==1.1.0
six==1.11.0
tb-nightly==1.15.0a20190806
tensorboard==1.10.0
tensorboardX==1.8
tensorflow==1.10.0
termcolor==1.1.0
torch==1.2.0
torchtext==0.4.0
torchvision==0.4.0a0+9232c4a
tqdm==4.30.0
urllib3==1.25.3
Werkzeug==0.15.4
